[squash] rename benchmarks to benchmark
