+++ test/parallel/test-benchmark-arrays.js
@@ -2,6 +2,6 @@
 
 require('../common');
 
-const runBenchmark = require('../common/benchmarks');
+const runBenchmark = require('../common/benchmark');
 
 runBenchmark('arrays', ['n=1', 'type=Array']);
