+++ test/parallel/test-benchmark-cluster.js
@@ -2,6 +2,6 @@
 
 require('../common');
 
-const runBenchmark = require('../common/benchmarks');
+const runBenchmark = require('../common/benchmark');
 
 runBenchmark('cluster', ['n=1', 'payload=string', 'sendsPerBroadcast=1']);
