+++ test/parallel/test-benchmark-domain.js
@@ -2,6 +2,6 @@
 
 require('../common');
 
-const runBenchmark = require('../common/benchmarks');
+const runBenchmark = require('../common/benchmark');
 
 runBenchmark('domain', ['n=1', 'arguments=0']);
