+++ test/parallel/test-benchmark-process.js
@@ -2,7 +2,7 @@
 
 require('../common');
 
-const runBenchmark = require('../common/benchmarks');
+const runBenchmark = require('../common/benchmark');
 
 runBenchmark('process',
              [
