+++ test/sequential/test-benchmark-http.js
@@ -9,7 +9,7 @@ if (!common.enoughTestMem)
 // rather than parallel to make sure it does not conflict with tests that choose
 // random available ports.
 
-const runBenchmark = require('../common/benchmarks');
+const runBenchmark = require('../common/benchmark');
 
 runBenchmark('http',
              [
