+++ test/sequential/test-benchmark-net.js
@@ -6,7 +6,7 @@ require('../common');
 // rather than parallel to make sure it does not conflict with tests that choose
 // random available ports.
 
-const runBenchmark = require('../common/benchmarks');
+const runBenchmark = require('../common/benchmark');
 
 runBenchmark('net',
              [
